requirements.txt
# Core web framework
fastapi
uvicorn

# Data models and validation
pydantic

# Machine Learning and AI


huggingface_hub
langchain
langchain_community
langchain_google_genai
langchain-huggingface
whisper


# Data processing
pandas
numpy
pypdf
docx2txt
pdfplumber

# Environment and configuration
python-dotenv

# File handling and storage
python-multipart
azure-storage-blob
boto3

# Communication services
twilio

# Audio processing
librosa

# Database
supabase

# Text processing
requests


# fastapi
# uvicorn
# pydantic
# faiss-cpu
# sentence-transformers
# huggingface_hub
# python-dotenv
# numpy

# langchain
# langchain_community
# langgraph
# langserve[all]
# langchain-ollama
# python-dotenv
# langchain_google_genai
# pandas
# sentence-transformers
# pypdf
# huggingface_hub
# langchain-huggingface
# whisper
# langchain-ollama
# cloudinary
# boto3==1.34.0
# python-multipart==0.0.6
# pdfplumber
# docx2txt
# twilio
# librosa
# supabase
# azure-storage-blob